+++ tomcat/src/main/java/org/apache/coyote/request/RequestUri.java
@@ -1,53 +1,21 @@
 package org.apache.coyote.request;
 
-import static java.util.stream.Collectors.*;
-
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
 public class RequestUri {
 
-	private static final String QUERY_STRING_SEPARATOR = "?";
-	private static final String QUERY_PARAM_DELIMITER = "&";
-	private static final String KEY_VALUE_SEPARATOR = "=";
-
 	private final String path;
 	private final Map<String, String> queryParams = new HashMap<>();
 
-	private RequestUri(final String path) {
+	public RequestUri(final String path) {
 		this(path, new HashMap<>());
 	}
 
-	private RequestUri(final String path, final Map<String, String> queryParams) {
+	public RequestUri(final String path, final Map<String, String> queryParams) {
 		this.path = path;
 		this.queryParams.putAll(queryParams);
-	}
-
-	public static RequestUri from(String uri) {
-		final var queryStringIndex = uri.indexOf(QUERY_STRING_SEPARATOR);
-
-		if (hasNoQuery(queryStringIndex)) {
-			return new RequestUri(uri);
-		}
-
-		final var path = uri.substring(0, queryStringIndex);
-		final var queryString = uri.substring(queryStringIndex + 1);
-		final var queryParams = parseQueryString(queryString);
-
-		return new RequestUri(path, queryParams);
-	}
-
-	private static Map<String, String> parseQueryString(final String queryString) {
-		return Arrays.stream(queryString.split(QUERY_PARAM_DELIMITER))
-			.map(queryParam -> queryParam.split(KEY_VALUE_SEPARATOR, 2))
-			.filter(parts -> parts.length == 2)
-			.collect(toMap(parts -> parts[0], parts -> parts[1]));
-	}
-
-	private static boolean hasNoQuery(final int index) {
-		return index == -1;
 	}
 
 	public boolean hasPath(final String path) {
